producer/requirements.txt
python-steam-api
beautifulsoup4
psycopg2-binary
kafka-python
pyTelegramBotAPI
requests
flask
